moving wrapped content fns into childSelectors
